Change last insertOrUpdateTime and LastPeekTime type from double to long.

+++ OutTimerSkipList.h
@@ -23,8 +23,8 @@ public:
     virtual int keysOnSpecificLevel(unsigned int level) const override;
     virtual int getMinKey() const override;
     virtual int getMaxLevel() override;
-    virtual double getLastInsertOrUpdateTime() const override;
-    virtual double getLastPeekTime() const override;
+    virtual long getLastInsertOrUpdateTime() const override;
+    virtual long getLastPeekTime() const override;
     virtual double getAverageInsertOrUpdateTime() const override;
     virtual double getAveragePeekTime() const override;
